requirements.txt
# Spark Core & SQL
pyspark==3.5.3

# Conectividad con Google Cloud
google-cloud-bigquery==3.10.0
google-cloud-storage==2.8.0

# Utilidades de Datos
pandas==2.2.3
numpy==1.26.4

# Pruebas y Calidad de Código
pytest==7.3.1
black==23.3.0